+++ _import/tycho-execution/src/encoding/evm/swap_encoder/swap_encoders.rs
@@ -43,8 +43,8 @@ impl SwapEncoder for UniswapV2SwapEncoder {
 
     fn encode_swap(
         &self,
-        swap: Swap,
-        encoding_context: EncodingContext,
+        swap: &Swap,
+        encoding_context: &EncodingContext,
     ) -> Result<Vec<u8>, EncodingError> {
         let token_in_address = bytes_to_address(&swap.token_in)?;
         let token_out_address = bytes_to_address(&swap.token_out)?;
@@ -101,8 +101,8 @@ impl SwapEncoder for UniswapV3SwapEncoder {
 
     fn encode_swap(
         &self,
-        swap: Swap,
-        encoding_context: EncodingContext,
+        swap: &Swap,
+        encoding_context: &EncodingContext,
     ) -> Result<Vec<u8>, EncodingError> {
         let token_in_address = bytes_to_address(&swap.token_in)?;
         let token_out_address = bytes_to_address(&swap.token_out)?;
@@ -110,7 +110,7 @@ impl SwapEncoder for UniswapV3SwapEncoder {
         let zero_to_one = Self::get_zero_to_one(token_in_address, token_out_address);
         let component_id = Address::from_str(&swap.component.id)
             .map_err(|_| EncodingError::FatalError("Invalid USV3 component id".to_string()))?;
-        let pool_fee_bytes = get_static_attribute(&swap, "fee")?;
+        let pool_fee_bytes = get_static_attribute(swap, "fee")?;
 
         let pool_fee_u24 = pad_to_fixed_size::<3>(&pool_fee_bytes)
             .map_err(|_| EncodingError::FatalError("Failed to extract fee bytes".to_string()))?;
@@ -162,15 +162,15 @@ impl SwapEncoder for UniswapV4SwapEncoder {
 
     fn encode_swap(
         &self,
-        swap: Swap,
-        encoding_context: EncodingContext,
+        swap: &Swap,
+        encoding_context: &EncodingContext,
     ) -> Result<Vec<u8>, EncodingError> {
-        let fee = get_static_attribute(&swap, "key_lp_fee")?;
+        let fee = get_static_attribute(swap, "key_lp_fee")?;
 
         let pool_fee_u24 = pad_to_fixed_size::<3>(&fee)
             .map_err(|_| EncodingError::FatalError("Failed to pad fee bytes".to_string()))?;
 
-        let tick_spacing = get_static_attribute(&swap, "tick_spacing")?;
+        let tick_spacing = get_static_attribute(swap, "tick_spacing")?;
 
         let pool_tick_spacing_u24 = pad_to_fixed_size::<3>(&tick_spacing).map_err(|_| {
             EncodingError::FatalError("Failed to pad tick spacing bytes".to_string())
@@ -245,15 +245,15 @@ impl SwapEncoder for BalancerV2SwapEncoder {
 
     fn encode_swap(
         &self,
-        swap: Swap,
-        encoding_context: EncodingContext,
+        swap: &Swap,
+        encoding_context: &EncodingContext,
     ) -> Result<Vec<u8>, EncodingError> {
         let token_approvals_manager = ProtocolApprovalsManager::new()?;
         let token = bytes_to_address(&swap.token_in)?;
         let approval_needed: bool;
 
-        if let Some(router_address) = encoding_context.router_address {
-            let tycho_router_address = bytes_to_address(&router_address)?;
+        if let Some(router_address) = &encoding_context.router_address {
+            let tycho_router_address = bytes_to_address(router_address)?;
             approval_needed = token_approvals_manager.approval_needed(
                 token,
                 tycho_router_address,
@@ -306,28 +306,28 @@ impl SwapEncoder for EkuboSwapEncoder {
 
     fn encode_swap(
         &self,
-        swap: Swap,
-        encoding_context: EncodingContext,
+        swap: &Swap,
+        encoding_context: &EncodingContext,
     ) -> Result<Vec<u8>, EncodingError> {
         if encoding_context.exact_out {
             return Err(EncodingError::InvalidInput("exact out swaps not implemented".to_string()));
         }
 
         let fee = u64::from_be_bytes(
-            get_static_attribute(&swap, "fee")?
+            get_static_attribute(swap, "fee")?
                 .try_into()
                 .map_err(|_| EncodingError::FatalError("fee should be an u64".to_string()))?,
         );
 
         let tick_spacing = u32::from_be_bytes(
-            get_static_attribute(&swap, "tick_spacing")?
+            get_static_attribute(swap, "tick_spacing")?
                 .try_into()
                 .map_err(|_| {
                     EncodingError::FatalError("tick_spacing should be an u32".to_string())
                 })?,
         );
 
-        let extension: Address = get_static_attribute(&swap, "extension")?
+        let extension: Address = get_static_attribute(swap, "extension")?
             .as_slice()
             .try_into()
             .map_err(|_| EncodingError::FatalError("extension should be an address".to_string()))?;
@@ -368,6 +368,7 @@ pub struct CurveSwapEncoder {
     executor_address: String,
     native_token_curve_address: String,
     native_token_address: Bytes,
+    wrapped_native_token_address: Bytes,
 }
 
 impl CurveSwapEncoder {
@@ -403,14 +404,32 @@ impl CurveSwapEncoder {
         }
     }
 
+    // Some curve pools support both ETH and WETH as tokens.
+    // They do the wrapping/unwrapping inside the pool
+    fn normalize_token(&self, token: Address, coins: &[Address]) -> Result<Address, EncodingError> {
+        let native_token_address = bytes_to_address(&self.native_token_address)?;
+        let wrapped_native_token_address = bytes_to_address(&self.wrapped_native_token_address)?;
+        if token == native_token_address && !coins.contains(&token) {
+            Ok(wrapped_native_token_address)
+        } else if token == wrapped_native_token_address && !coins.contains(&token) {
+            Ok(native_token_address)
+        } else {
+            Ok(token)
+        }
+    }
+
     fn get_coin_indexes(
         &self,
-        swap: Swap,
+        swap: &Swap,
         token_in: Address,
         token_out: Address,
     ) -> Result<(U8, U8), EncodingError> {
-        let coins_bytes = get_static_attribute(&swap, "coins")?;
+        let coins_bytes = get_static_attribute(swap, "coins")?;
         let coins: Vec<Address> = from_str(std::str::from_utf8(&coins_bytes)?)?;
+
+        let token_in = self.normalize_token(token_in, &coins)?;
+        let token_out = self.normalize_token(token_out, &coins)?;
+
         let i = coins
             .iter()
             .position(|&addr| addr == token_in)
@@ -446,13 +465,14 @@ impl SwapEncoder for CurveSwapEncoder {
             executor_address,
             native_token_address: chain.native_token()?,
             native_token_curve_address,
+            wrapped_native_token_address: chain.wrapped_token()?,
         })
     }
 
     fn encode_swap(
         &self,
-        swap: Swap,
-        encoding_context: EncodingContext,
+        swap: &Swap,
+        encoding_context: &EncodingContext,
     ) -> Result<Vec<u8>, EncodingError> {
         let token_approvals_manager = ProtocolApprovalsManager::new()?;
         let native_token_curve_address = Address::from_str(&self.native_token_curve_address)
@@ -473,9 +493,9 @@ impl SwapEncoder for CurveSwapEncoder {
 
         let component_address = Address::from_str(&swap.component.id)
             .map_err(|_| EncodingError::FatalError("Invalid curve pool address".to_string()))?;
-        if let Some(router_address) = encoding_context.router_address {
+        if let Some(router_address) = &encoding_context.router_address {
             if token_in != native_token_curve_address {
-                let tycho_router_address = bytes_to_address(&router_address)?;
+                let tycho_router_address = bytes_to_address(router_address)?;
                 approval_needed = token_approvals_manager.approval_needed(
                     token_in,
                     tycho_router_address,
@@ -488,7 +508,7 @@ impl SwapEncoder for CurveSwapEncoder {
             approval_needed = true;
         }
 
-        let factory_bytes = get_static_attribute(&swap, "factory")?.to_vec();
+        let factory_bytes = get_static_attribute(swap, "factory")?.to_vec();
         // the conversion to Address is necessary to checksum the address
         let factory_address =
             Address::from_str(std::str::from_utf8(&factory_bytes).map_err(|_| {
@@ -548,8 +568,8 @@ impl SwapEncoder for MaverickV2SwapEncoder {
 
     fn encode_swap(
         &self,
-        swap: Swap,
-        encoding_context: EncodingContext,
+        swap: &Swap,
+        encoding_context: &EncodingContext,
     ) -> Result<Vec<u8>, EncodingError> {
         let component_id = AlloyBytes::from_str(&swap.component.id)
             .map_err(|_| EncodingError::FatalError("Invalid component ID".to_string()))?;
@@ -591,8 +611,8 @@ impl SwapEncoder for BalancerV3SwapEncoder {
 
     fn encode_swap(
         &self,
-        swap: Swap,
-        encoding_context: EncodingContext,
+        swap: &Swap,
+        encoding_context: &EncodingContext,
     ) -> Result<Vec<u8>, EncodingError> {
         let pool = Address::from_str(&swap.component.id).map_err(|_| {
             EncodingError::FatalError("Invalid pool address for Balancer v3".to_string())
@@ -663,7 +683,7 @@ mod tests {
             )
             .unwrap();
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);
             assert_eq!(
@@ -723,7 +743,7 @@ mod tests {
             )
             .unwrap();
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);
             assert_eq!(
@@ -788,7 +808,7 @@ mod tests {
             )
             .unwrap();
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);
 
@@ -860,7 +880,7 @@ mod tests {
             )
             .unwrap();
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);
 
@@ -933,7 +953,7 @@ mod tests {
             )
             .unwrap();
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);
 
@@ -1028,10 +1048,10 @@ mod tests {
             )
             .unwrap();
             let initial_encoded_swap = encoder
-                .encode_swap(initial_swap, context.clone())
+                .encode_swap(&initial_swap, &context)
                 .unwrap();
             let second_encoded_swap = encoder
-                .encode_swap(second_swap, context)
+                .encode_swap(&second_swap, &context)
                 .unwrap();
 
             let combined_hex =
@@ -1112,7 +1132,7 @@ mod tests {
                     .unwrap();
 
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
 
             let hex_swap = encode(&encoded_swap);
@@ -1188,11 +1208,11 @@ mod tests {
             };
 
             let first_encoded_swap = encoder
-                .encode_swap(first_swap, encoding_context.clone())
+                .encode_swap(&first_swap, &encoding_context)
                 .unwrap();
 
             let second_encoded_swap = encoder
-                .encode_swap(second_swap, encoding_context)
+                .encode_swap(&second_swap, &encoding_context)
                 .unwrap();
 
             let combined_hex =
@@ -1314,7 +1334,7 @@ mod tests {
             .unwrap();
             let (i, j) = encoder
                 .get_coin_indexes(
-                    swap,
+                    &swap,
                     Address::from_str(token_in).unwrap(),
                     Address::from_str(token_out).unwrap(),
                 )
@@ -1366,7 +1386,7 @@ mod tests {
             )
             .unwrap();
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);
 
@@ -1438,7 +1458,7 @@ mod tests {
             )
             .unwrap();
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);
 
@@ -1520,7 +1540,7 @@ mod tests {
             )
             .unwrap();
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);
 
@@ -1585,7 +1605,7 @@ mod tests {
             )
             .unwrap();
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);
 
@@ -1644,7 +1664,7 @@ mod tests {
             .unwrap();
 
             let encoded_swap = encoder
-                .encode_swap(swap, encoding_context)
+                .encode_swap(&swap, &encoding_context)
                 .unwrap();
             let hex_swap = encode(&encoded_swap);